MLR-229: Refacotred workflow services into their own java package to seperate them from client-based services. Added clients for the CRU and File Export services. Cleaned up the workflow services code to use the newly created services. Renamed and moved some static variables around. Removed duplicate definitions of some static variables. Added additional logging of some errors.
